+++ src/types/FileMakerDataAPIResponse/FileMakerDataAPIGetRecordRangeResponse.ts
@@ -1,11 +1,12 @@
+import FileMakerDataAPIRecord from '../FileMakerDataAPIRecord'
 import FileMakerDataAPIGetRecordResponse from './FileMakerDataAPIGetRecordResponse'
 
 /**
  * Represents the response from the FileMaker Data API when getting a range of records.
  */
 type FileMakerDataAPIGetRecordRangeResponse<
-  FieldData = Record<string, unknown>,
-  PortalData = Record<string, unknown>
+  FieldData = FileMakerDataAPIRecord['fieldData'],
+  PortalData = FileMakerDataAPIRecord['portalData']
 > = FileMakerDataAPIGetRecordResponse<FieldData, PortalData>
 
 export default FileMakerDataAPIGetRecordRangeResponse